Create new render-specific structs; improve tag render
Structs are created just for sending to Handlebars renderer, and can derive some info automatically so it doesn't
need to be explicitly passed to the constructor.

Also, it seems Handlebars was taking quite some time to render tags; pages now render
twice as fast(!) using a simple string concatenation helper instead of a template file.