Major refactor of article build/store/view
Renamed ContentView to ParsedArticle and moved it into the builder
module. Storage module is now only concerned with fetching and updating
articles; builder is responsible for reading from disk and parsing. View
module is simplified to just index and article views, and in all cases,
everything is borrowed from the master ParsedArticle vec provided by
storage, which means a lot fewer clones.